fix: auto loading post.|feature: check update.
